Throw exception if end is before start in SlotServiceImpl
The slot calculation only works, if the end date is equal to or after the start date. As a consequence, an empty list has been returned. This could potentially trigger strange behavior within the application, if start and end date are mixed up (as happened in the broken test).

To make this mistake obvious, I added an IllegalArgumentException if the end date is before the start date.

I also changed the test to use LocalDate instead of LocalDateTime, to avoid further confusion with time stamps.