#104 Refactor location based search to be performed in Java
This is a large rework of how location based searches are performed. Until now, the database query calculated a rough estimate of the distance to the shop for all shops in the shop table before fetching them, regardless of whether they are within a desired max distance to the user or not. We are currently not performing location based searches in the frontend, but this is subject to change.

The new behavior is as follows: if you don't restrict the search/find (as the frontend does at the moment) you will get all active shops with a fast query which doesn't do any calculation.

If you want to restrict the query to a certain max distance from the user, you can do so using the optional maxDistance query parameter in both resources (search and nearby). This will calculate a bounding box around the given location and perform a fast where query without calculation to get all shops within that box. The shop service then does the actual filtering on the shops within that box. This should be substantially faster than calculating the distances for all shops in the DB beforehand.