#189 Fix incorrectly filtered out cities in the location service
We deduplicated the results of two queries based on a ZIP code, which is not a unique identifier and therefore involuntarily removed cities which share their ZIP code with another city.

We now also return all matches in case of a complete ZIP code instead of limiting the result to 5 results, since we might have more than 5 cities with the same zip code.