Add LogoutInterceptor
This will detect 401 responses on all returning HttpClient requests and redirects the user accordingly.

Admins will see the login page for admins, shop owners the regular login page. The original "login" API calls, however, are not triggering this redirect (because we're doing these requests at the very startup of the app to determine whether we're having a session right now or not)