[#230] Add slug to shops
This was a bit of a hassle, as we need to generate slugs for existing shops. I didn't want to do that in SQL, so i wrote a Java based Flyway migration. But there we can't inject Spring beans, so I had to refactor the SlugService to keep dependencies on Spring magic to a minimum